@@ -2,12 +2,12 @@
     "type": "project",
     "license": "proprietary",
     "require": {
-        "php": ">=7.1.3",
+        "php": ">=8.0",
         "ext-ctype": "*",
         "ext-iconv": "*",
         "composer/package-versions-deprecated": "1.11.99.1",
         "doctrine/annotations": "^1.0",
-        "doctrine/doctrine-bundle": "^2.2",
+        "doctrine/doctrine-bundle": "^2.11",
         "doctrine/doctrine-fixtures-bundle": "^3.4",
         "doctrine/doctrine-migrations-bundle": "^3.0",
         "doctrine/orm": "^2.8",
@@ -40,6 +40,7 @@
         "twig/twig": "^2.12|^3.0"
     },
     "require-dev": {
+        "rector/rector": "^0.19.0",
         "symfony/browser-kit": "^6.0",
         "symfony/css-selector": "^6.0",
         "symfony/debug-bundle": "^6.0",
@@ -61,7 +62,7 @@
     },
     "autoload": {
         "psr-4": {
-            "App\\": [ "altermeso/src/", "src/" ]
+            "App\\": [ "src/", "src/" ]
         }
     },
     "autoload-dev": {
